feat(output and puzzle): new classes Output and Puzzle, with function for work.Solution => Redeem
Output represents a Boost output, including prevout.

Puzzle represents a Boost output with a private key
has been assigned to it, which may have happened
before or after the output was created, depending on
whether it is Bounty or Contract. It has a method
called redeem, which takes a work.Solution and an
incomplete transaction and returns a Redeem object.
In effect it creates the signature for the redeem
script, since everything else about it can be known
from the hash puzzle solution.

There is a test that fails due to an exception that
gets thrown and I'm not sure why. e